Move pirate_channel_init() and pirate_channel_term() out of libpirate
Want to keep C constructor and destructor as an example of how
channels could be initialized. To be determined whether this
functionality should live in the pirate primitives, or this
functionality should be inserted by the GAPS compiler toolchain,
or initialization and termination should be explicit operations
(not using the constructor).

Moved pirate_channel_init() and pirate_channel_term() into
initialize.c. Not included in libpirate compilation.

Removed perror() and printf() messages from initialize.c.
To be determined whether libpirate should print messages.