UDP channel store previous write contents
Store the previous send() request for the UDP channel. On receiving
an ECONNREFUSED error when attempting a send(), then resend the
previous packet from the buffer and then resend the current
packet.

ECONNREFUSED are generated on even-numbered sends. On opening
a UDP channel for writing send a zero byte UDP packet. This
ensures that the first user pirate_write() will retry on
ECONNREFUSED from the zero byte packet. On opening a UDP channel
for reading then the first read silently consumes the zero
byte packet.

Tested successfully on the libpirate unit tests, the channel
demo, and the pnt demo. channel and pnt demos with udp channels
become deterministic with this patch and behavior is invariant
to the ordering and timing of starting the processes.

Not implemented for GE channel types. Need to determine whether
ICMP Destination Unreachable packets are delivered in that environment.
Also need to determine whether zero byte packets can be sent
over GE channels without being filtered. For GE channels are either
of these permissible: packets wih no data and no gaps metadata,
packets with no data and with gaps metadata?