Harmonize close() semantics on unix pipe, unix socket, and tcp socket
If the writer closes first then the reader receives a return value
of 0 on the next read() request. If the reader closes first then
the writer receives a SIGPIPE signal and an EPIPE errno on the next
write() request. Add one set of shared unit tests to confirm same
behavior across unix pipe, unix socket, and tcp socket.

On the tcp reader set socket option SO_LINGER to onoff=1 and linger=0.
This causes the reader to send a RST packet and reset the connection
on close(). Graceful TCP connection termination is two pairs of two-way
FIN, ACK handshakes. Graceful termination would require the writer to
read() and receive a return value of 0. Since a gaps writer must not
call read() we use a RST packet sent from the reader to close the
conection. libpirate ignores the ECONNRESET error code when received
by the tcp writer on close(). SO_LINGER socket option is not set on the
writer.