camera demo manual partition (#140)
Manually partition the camera demo so that input devices and the output device communicate over gaps channels. Updated the camera demo so that all of the components are now optional and disabled by default: video source, input devices, output device, and frame processors. This allows for an executable to be started that only functions as an input device, or only functions as a video source and an output device, for example.

Added basic error detection of missing gaps channels. If a component requires a request channel or a response channel and that channel has not been specified then an error is reported. Some errors cannot be detected. If you start two instances as input devices but the output device has only opened one gaps request channel then one of the input devices will not communicate with the output device.

Updated README.md with new component system.