idl code generation bug fixes
Fixed a bug in code generation for enums inside unions with
C language target. The generated code included a namespace
prefix and C enums do not have namespaces.

(I believe) fixed a bug in the code generation for enums inside
unums with C++ language target. The generated code included
the namespace prefix but not the enum typename prefix. Modified
the code generation to include the namespace prefix and the
enum typename prefix.

Added an enum as the tag type to nested.idl example to test
more complex examples.

Fixed DFDL language target. Split the declaration
of the xml simpleType/complexType and the xml element.
Otherwise types cannot be reused.

Added test/output/dfdl with an example of generated code.