camera demo: support MPEG-TS video stream reconnection
The ffmpeg library does not support a separate timeout
variable for open() and read() operations on UDP video
streams. Therefore we teardown and reconnect the video
stream when read() encounters a timeout.

Successfully reproduced the error condition using the
Raspbery Pi camera. Verified that the patch fixes the
behavior. Ran the patch under valgrind to test for
memory leaks (see below).

This patch also reverts the use of ptrace() to determine
whether or not we are running under gdb. The behavior
doesn't work under valgrind and prevents us from using
valgrind. Reverted to a manual command-line option that
can be enabled when running under gdb.