vscode plug-in: fix tsconfig and reorganize projects
The current setup is deceptively not working: VSCode does not pick up on
tsconfig-webview.json due to a long-standing bug where it only considers files
named tsconfig.json.

I suggest using the following setup, where a common 'tsconfig-default.json'
contains all the common definitions, and each subdirectory points and extends
it.  The top-level 'tsconfig.json' references the important sub-projects.

This is the setup used in the official TypeScript repository, so it seems like
it will be well supported.  Note that it requires running:

tsc -b ./path/to/some/tsconfig.json

Rather than using the '-p' option.  This is updated in 'package.json' so it
should not matter if you use 'npm run' commands.

Finally, I also opted to explicitly list out the files, since there are only a
few, and it avoids compile problems when switching between branches while some
untracked source files are present in the source directory.