vscode plug-in: display connections and update listeners
This commit adds basic display of connections.  Because connections coordinates
are relative to that of their endpoints, this required a modification to the
listeners protocol to allow registering multiple listener to the same tracked
identifiers.

This allows separating some concerns, as the connection is responsible for
registering itself as interested in changes to its endpoints, rather than
having the endpoints be concerned with broadcasting their changes to an unknown
set of other components who depend on those values.

We also now send updates to the view who made those changes, since other
components in the same view will be listening.  This actually simplifies the
sender of those changes, as they don't need to do their modification in place
and can listen to the very changes they emit.

There is a bit of code duplication / lack of abstraction in the handling of
endpoints, as they can be either buses or actor ports and we treat those very
differently.  This can likely be improved down the line.

The display is subject to change in future commits: in particular, we will
likely want the line to be displayed under the port, and we will also want the
line to slant if an actor is too far from the bus it is connected to.