vscode plug-in: rearchitecture
This commit reimplements the VSCode extension using pure views (built using
React hooks).

The previous changes made it so that the same piece of model data was sometimes
located in two different locations, which created all sorts of synchronization
issues.

With this new architecture, the webview component is solely responsible for
changing the model, and synchronizing the model changes back and forth with the
extension backend.

All views receive the relevant parts of the state as input, and receive
handlers for asking the webview to change their state (or any part of the
state).  The webview is then responsible for updating its state, as well as
propagating the changes to the backend, and then it re-renders the views with
the new state.

With this, we no longer need the multiple listeners introduced in a previous
commit, as the webview has, as it used to be, exactly one listener per tracked
field.  Dependent views, like the "connection" view, receive all the data they
need to compute their position from the webview.

We also no longer need system services as an abstraction, as all of the
responsibilities it was passing down to views are now handled within the
top-level webview component.

The parser was also updated to give us locations for entire objects like
actors, buses, components, such that it is now feasible (though not completely
implemented) to delete them from the webview easily.

Adding them from the webview would be slightly less tricky, as they would need
to be appended to the document, but it sounds pretty easy with the updated
architecture.

The big benefit of this change is that we may no longer run into issues of
stale view objects being present in the view were we to delete them from the
model: as the view is a pure function of the model data, when an object is
changed or removed in the model, all of the DOM/SVG elements that were part of
it are automatically removed from the view.